requirements.txt
pandas>=2.2.3
plotly>=5.24.1
psycopg2-binary>=2.9.10
requests>=2.32.3
streamlit>=1.39.0
